requirements.txt
google-generativeai>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.14.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

